refact/front/BtnHomeGreen: moves to src/components/buttons/BtnHomeGreen.jsx
